refactor(#302): split ERD mutation target resolution
