+++ apps/backend/core/src/main/java/com/schemafy/core/erd/operation/application/service/FinalizedErdMutationTarget.java
@@ -1,0 +1,10 @@
+package com.schemafy.core.erd.operation.application.service;
+
+import com.schemafy.core.erd.operation.domain.ErdTouchedEntity;
+
+record FinalizedErdMutationTarget(
+    String projectId,
+    String schemaId,
+    ErdTouchedEntity touchedEntity) {
+
+}
